Restructure initialization code and root datastructure
This commit changes the initialization code and data layout so that all
data stored in persistent memory hangs off the root pointer. Earlier,
the log data was stored in a region of memory carved out from the
persitent memory region that is not managed by the runtime. This region
was manually divided into header blocks and linked to the transaction
library at run time.